requirements.txt
fastapi
uvicorn
httpx
pyyaml
pydantic